root/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "olrut", email = "121128961+olrut@users.noreply.github.com" }
]
dependencies = [
    "fastapi>=0.115.4,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "fastapi-users[sqlalchemy]>=14.0.0,<15.0.0",
    "fastapi-users-db-sqlalchemy>=6.0.0,<7.0.0",
    "config>=0.5.1,<0.6.0",
    "requests>=2.32.3,<3.0.0",
    "PyJWT>=2.9.0,<3.0.0",
    
    "httpx-oauth>=0.15.1,<0.16.0",
    "python-dotenv>=1.0.1,<2.0.0"
]

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "main.py", "README.md", "pyproject.toml"]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "aiosqlite>=0.19.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
]
